requirements.txt
aiofiles==24.1.0
aiolimiter==1.1.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
autograd==1.6.2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.21.0
beartype==0.18.5
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cryptography==43.0.0
cycler==0.12.1
dask==2024.7.1
dask-expr==1.1.9
datashaper==0.0.49
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
environs==11.0.0
exceptiongroup==1.2.2
executing==2.0.1
fastparquet==2024.5.0
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
graphrag==0.2.0
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyppo==0.4.0
idna==3.7
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lancedb==0.10.2
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.37.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
POT==0.9.4
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.14.1
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.37.0
swifter==1.4.0
tenacity==8.5.0
textual==0.72.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==2.2.2
uvloop==0.19.0
wcwidth==0.2.13
widgetsnbextension==4.0.11
wrapt==1.16.0
yfiles_jupyter_graphs==1.7.3
zipp==3.19.2
